MIM-1996:Refaktorering av kode
Flyttet kode til utils, kan brukes av attachmentTablesFigures og etterhvert
